requirements.txt
# Core dependencies
pymilvus==2.3.4
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==5.2.0
ollama==0.1.6

# Document processing
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.2

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
tqdm==4.66.1

# Optional: For advanced features
chromadb==0.4.22